+++ src/pal.ts
@@ -1,6 +1,8 @@
 import { NetworkClientFactory } from "@iota-pico/core/dist/factories/networkClientFactory";
+import { RngServiceFactory } from "@iota-pico/core/dist/factories/rngServiceFactory";
 import { INetworkEndPoint } from "@iota-pico/core/dist/interfaces/INetworkEndPoint";
 import { NetworkClient } from "./network/networkClient";
+import { RngService } from "./services/rngService";
 
 /**
  * Platform abstraction layer for NodeJS.
@@ -16,6 +18,7 @@ export class PAL {
         if (!PAL._isInitialized) {
             PAL._isInitialized = true;
             NetworkClientFactory.instance().register("default", (networkEndpoint: INetworkEndPoint) => new NetworkClient(networkEndpoint));
+            RngServiceFactory.instance().register("default", () => new RngService());
         }
 
         return Promise.resolve();