+++ test/pal.spec.ts
@@ -2,6 +2,7 @@
  * Tests for PAL.
  */
 import { NetworkClientFactory } from "@iota-pico/core/dist/factories/networkClientFactory";
+import { RngServiceFactory } from "@iota-pico/core/dist/factories/rngServiceFactory";
 import { NetworkEndPoint } from "@iota-pico/core/dist/network/networkEndPoint";
 import * as chai from "chai";
 import { PAL } from "../src/pal";
@@ -16,6 +17,11 @@ describe("PAL", () => {
         it("can be called and register network client", async () => {
             await PAL.initialize();
             chai.expect(NetworkClientFactory.instance().create("default", new NetworkEndPoint("http", "localhost", 14265))).to.not.be.equal(undefined);
+        });
+
+        it("can be called and register rng service", async () => {
+            await PAL.initialize();
+            chai.expect(RngServiceFactory.instance().create("default")).to.not.be.equal(undefined);
         });
 
         it("can be called twice", async () => {